I give up, this is becoming a repo'd project
This is some of the most hideous, if not the most hideous, code I've
ever written. At this point, only an "admin" interface is partially
enabled, you have to fake serial input, and there's no main game
display.

BUT.

It's fully functional as far as standard Jeopardy questions are
concerned. And I have been told that we only care about that for now.

So.

Here it is, what is currently implemented. Next stop? Game display.