More performant fix inversion of the last commit
The problem was that we were waiting around to calculate the round
scores, no matter what (and I do know that the round scores are
calculated in the worst way possible), just because a paint happened
before a score update.

Now, the score update happens before the paint. However, it's imperative
that it happens after the verification of whether the contestant got the
question or not, so it actually means something. This is why the same
update is run in both branches of the loop: it's a poorly structured
loop, with a side-effectful evaluation, and the following lines in both
branches really need it.

Geez, I'm tired. It's becoming clear in the code.